Move Block_Trait into the Traits directory
