+++ src/Tribe/Editor/Blocks/Event_Links.php
@@ -1,7 +1,7 @@
 <?php
 class Tribe__Events__Editor__Blocks__Event_Links
 extends Tribe__Editor__Blocks__Abstract {
-	use Tribe\Events\Editor\Blocks\Block_Trait;
+	use Tribe\Events\Traits\Block_Trait;
 
 	/**
 	 * Which is the name/slug of this block
